Refactors the logic inside win.webContents.executeJavaScript by moving it to another file and then exporting the module.
